Fix Socket.IO loading issue in driver interface
The local socket.io.min.js file was causing a syntax error preventing
the driver interface from loading jobs.

Solution: Replaced the local file with a CDN-hosted version of Socket.IO
client 4.5.4, which is compatible with the server version and guaranteed
to be valid.

This will fix the 'io is not defined' error and allow the driver
interface to load available jobs and deliveries.